Moved DefaultDocumentLayout from /core to /default_editor. Added TapSequenceGestureRecognizer to /infrastructure.
