+++ lib/flutter_richtext.dart
@@ -3,3 +3,11 @@ library flutter_richtext;
 export 'src/infrastructure/selectable_text.dart';
 export 'src/infrastructure/attributed_text.dart';
 export 'src/infrastructure/attributed_spans.dart';
+export 'src/infrastructure/multi_tap_gesture.dart';
+export 'src/core/document.dart';
+export 'src/core/document_composer.dart';
+export 'src/core/document_editor.dart';
+export 'src/core/document_layout.dart';
+export 'src/core/document_selection.dart';
+export 'src/core/edit_context.dart';
+export 'src/default_editor/layout.dart';
